+++ RELEASES.md
@@ -1,3 +1,10 @@
+# Astra Monitor 15 - TO BE RELEASED
+
+### Bug fixes
+- Fixed too wide Select Box in Preferences [[#80](https://github.com/AstraExt/astra-monitor/issues/80)]
+- Fixed CPU Cores Usage Info not being updated when Cores Bar Breakdown is disabled [[#83](https://github.com/AstraExt/astra-monitor/issues/83)]
+
+
 # Astra Monitor 14 - March 11 2024
 
 ### Hwmon